move volumetric messages into api folder
