+++ nanome/api/message_serializers.py
@@ -2,9 +2,8 @@ import logging
 
 from nanome._internal import serializer_fields as serializer_fields
 from nanome._internal.enums import IntegrationCommands
-from nanome.api import integration, structure, volumetric
+from nanome.api import integration
 from nanome.api._hashes import Hashes
-from nanome.api.user import PresenterInfo
 
 
 logger = logging.getLogger(__name__)
@@ -229,52 +228,3 @@ class GetControllerTransformsResponse(serializer_fields.TypeSerializer):
                   left_controller_rotation, right_controller_position, right_controller_rotation)
         return result
 
-
-class AddVolume(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        self.__complex = structure.serializers.ComplexSerializer()
-        atom_serializer = structure.serializers.AtomSerializer()
-        long_serializer = serializer_fields.LongSerializer()
-        self.__dict = serializer_fields.DictionarySerializer()
-        self.__dict.set_types(long_serializer, atom_serializer)
-        self.__data = volumetric.serializers.VolumeDataSerializer()
-        self.__properties = volumetric.serializers.VolumePropertiesSerializer()
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "AddVolume"
-
-    def serialize(self, version, value, context):
-        subcontext = context.create_sub_context()
-        subcontext.payload["Atom"] = {}
-        subcontext.write_using_serializer(self.__complex, value[0])
-        context.write_using_serializer(self.__dict, subcontext.payload["Atom"])
-        context.write_bytes(subcontext.to_array())
-
-        context.write_long(value[1])
-        context.write_using_serializer(self.__data, value[2])
-        context.write_using_serializer(self.__properties, value[3])
-
-    def deserialize(self, version, context):
-        raise NotImplementedError
-
-
-class AddVolumeDone(serializer_fields.TypeSerializer):
-
-    def __init__(self):
-        pass
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "AddVolumeDone"
-
-    def serialize(self, version, value, context):
-        raise NotImplementedError
-
-    def deserialize(self, version, context):
-        return None
