+++ nanome/api/volumetric/__init__.py
@@ -2,4 +2,4 @@ from .models import UnitCell, VolumeData, VolumeLayer, VolumeProperties
 from . import serializers
 
 # folders
-from . import io
+from . import io, messages
